Drop mkdocs build path
- Delete mkdocs.yml and scripts/copy_docs_md.py.
- Drop mkdocs / mkdocs-material / mkdocs-jupyter / mkdocs-section-index /
  mkdocs-include-markdown-plugin / mkdocs-git-revision-date-localized-plugin /
  mike from the docs extras; mkdocstrings[python] stays for zensical's compat
  layer.
- Inline the HTML→markdown helpers from copy_docs_md.py into post_zensical.py
  so it's self-contained.
- Replace the mkdocs Makefile targets with `docs` / `docs-serve` calling the
  zensical pipeline.
