--- mkdocs.yml
@@ -1,127 +1,0 @@
-site_name: wigglystuff
-site_description: Creative AnyWidgets for notebooks
-site_url: https://wigglystuff.dev
-repo_url: https://github.com/koaning/wigglystuff
-repo_name: koaning/wigglystuff
-docs_dir: mkdocs
-site_dir: site
-theme:
-  name: material
-  custom_dir: mkdocs/overrides
-  logo: assets/logo.png
-  favicon: assets/logo.png
-  icon:
-    repo: fontawesome/brands/github
-  features:
-    - navigation.tabs
-    - navigation.sections
-    - navigation.expand
-    - navigation.path
-    - toc.follow
-    - content.tabs.link
-    - content.code.annotate
-    - search.suggest
-    - search.highlight
-    - navigation.instant
-  palette:
-    scheme: default
-    primary: grey
-    accent: blue grey
-  font:
-    text: "PT Sans"
-    code: "Fira Mono"
-nav:
-  - Overview: index.md
-  - Reference:
-      - Overview: reference/index.md
-      - Slider2D: reference/slider2d.md
-      - ChartPuck: reference/chart-puck.md
-      - ChartMultiSelect: reference/chart-multi-select.md
-      - ChartSelect: reference/chart-select.md
-      - ParallelCoordinates: reference/parallel-coords.md
-      - Matrix: reference/matrix.md
-      - SortableList: reference/sortable-list.md
-      - Paint: reference/paint.md
-      - ThreeWidget: reference/three-widget.md
-      - EdgeDraw: reference/edge-draw.md
-      - KeystrokeWidget: reference/keystroke.md
-      - WebkitSpeechToText: reference/talk.md
-      - ColorPicker: reference/color-picker.md
-      - CopyToClipboard: reference/copy-to-clipboard.md
-      - Tangle Widgets: reference/tangle.md
-      - GamepadWidget: reference/gamepad.md
-      - WebcamCapture: reference/webcam-capture.md
-      - CellTour: reference/cell-tour.md
-      - ImageRefreshWidget: reference/image-refresh.md
-      - HTMLRefreshWidget: reference/html-refresh.md
-      - ProgressBar: reference/progress-bar.md
-      - PulsarChart: reference/pulsar-chart.md
-      - ModuleTreeWidget: reference/module-tree.md
-      - Neo4jWidget: reference/neo4j-widget.md
-      - ScatterWidget: reference/scatter-widget.md
-      - DiffViewer: reference/diff-viewer.md
-      - SplineDraw: reference/spline-draw.md
-      - TextCompare: reference/text-compare.md
-      - EnvConfig: reference/env-config.md
-      - AltairWidget: reference/altair-widget.md
-      - ApiDoc: reference/api-doc.md
-      - PlaySlider: reference/play-slider.md
-      - HoverZoom: reference/hover-zoom.md
-      - Treemap: reference/treemap.md
-      - NestedTable: reference/nested-table.md
-      - Utils: reference/utils.md
-plugins:
-  - search
-  - section-index
-  - mkdocs-jupyter:
-      execute: false
-      include_source: true
-  - mkdocstrings:
-      handlers:
-        python:
-          paths: ["."]
-          options:
-            docstring_style: google
-            merge_init_into_class: true
-            show_source: true
-            show_root_heading: false
-            show_signature: true
-            show_signature_annotations: true
-            signature_crossrefs: true
-            separate_signature: true
-            heading_level: 2
-markdown_extensions:
-  - admonition
-  - attr_list
-  - def_list
-  - footnotes
-  - md_in_html
-  - toc:
-      permalink: "#"
-  - pymdownx.details
-  - pymdownx.snippets
-  - pymdownx.highlight:
-      anchor_linenums: true
-      pygments_lang_class: true
-  - pymdownx.inlinehilite
-  - pymdownx.magiclink
-  - pymdownx.superfences
-  - pymdownx.tabbed:
-      alternate_style: true
-  - pymdownx.tasklist:
-      custom_checkbox: true
-  - pymdownx.emoji:
-      emoji_generator: !!python/name:pymdownx.emoji.to_svg
-extra_css:
-  - assets/stylesheets/extra.css
-extra:
-  social:
-    - icon: fontawesome/brands/github
-      link: https://github.com/vdamario/wigglystuff
-      name: GitHub
-    - icon: material/notebook
-      link: https://marimo.io/gallery
-      name: Marimo gallery
-watch:
-  - mkdocs.yml
-  - mkdocs
