+++ pyproject.toml
@@ -47,12 +47,6 @@ docs = [
     "altair>=6.0.0",
     "black>=24.8.0",
     "marimo>=0.18.0",
-    "mike>=2.1.0",
-    "mkdocs-git-revision-date-localized-plugin>=1.2.6",
-    "mkdocs-include-markdown-plugin>=6.2.1",
-    "mkdocs-jupyter>=0.25.0",
-    "mkdocs-material>=9.5.0",
-    "mkdocs-section-index>=0.3.6",
     "mkdocstrings[python]>=0.25.1",
     "mohtml>=0.1.11",
     "pandas>=2.3.3",
